test: split tests according to typestate
Multi-line description of commit,
feel free to be detailed.

[Ticket: X]
